RGR-Andrey/realty_agency/requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
pymongo>=4.6.1
redis>=5.0.1
